+++ packages/cross-chain-core/src/CrossChainCore.ts
@@ -23,6 +23,7 @@ import {
   getAptosWalletUSDCBalance,
   getEthereumWalletUSDCBalance,
   getSolanaWalletUSDCBalance,
+  getSuiWalletUSDCBalance,
 } from "./utils/getUsdcBalance";
 
 export interface CrossChainDappConfig {
@@ -54,7 +55,8 @@ export type Chain =
   | "Base"
   | "Arbitrum"
   | "PolygonSepolia"
-  | "Polygon";
+  | "Polygon"
+  | "Sui";
 
 // Map of Ethereum chain id to testnet chain config
 export const EthereumChainIdToTestnetChain: Record<string, ChainConfig> = {
@@ -134,6 +136,7 @@ export class CrossChainCore {
         this._dappConfig.aptosNetwork,
       );
     }
+
     if (!this.CHAINS[sourceChain]) {
       throw new Error(`Unsupported chain: ${sourceChain}`);
     }
@@ -159,6 +162,12 @@ export class CrossChainCore {
           this._dappConfig.aptosNetwork,
           sourceChain,
           // TODO: maybe let the user config it
+          this.CHAINS[sourceChain].defaultRpc,
+        );
+      case "Sui":
+        return await getSuiWalletUSDCBalance(
+          walletAddress,
+          this._dappConfig.aptosNetwork,
           this.CHAINS[sourceChain].defaultRpc,
         );
       default: