+++ packages/derived-wallet-sui/src/index.ts
@@ -1,0 +1,5 @@
+export * from "./SuiDerivedPublicKey";
+export * from "./SuiDerivedWallet";
+export * from "./setupAutomaticDerivation";
+export * from "./shared";
+export { fromBase64, fromHex, toBase64, toHex } from "@mysten/sui/utils";