+++ packages/wallet-adapter-core/src/constants.ts
@@ -29,6 +29,17 @@ export const APTOS_CONNECT_BASE_URL = "https://aptosconnect.app";
 /** The base URL for all Petra Web wallets. */
 export const PETRA_WEB_BASE_URL = "https://web.petra.app";
 
+/** The name of the generic wallet for Petra Web. */
+export const PETRA_WEB_GENERIC_WALLET_NAME = "Petra Web";
+
+/** The name of the Petra wallet. */
+export const PETRA_WALLET_NAME = "Petra";
+
+/** The default connection fallbacks for wallets that are not installed. */
+export const DEFAULT_WALLET_CONNECTION_FALLBACKS = {
+  [PETRA_WALLET_NAME]: PETRA_WEB_GENERIC_WALLET_NAME,
+};
+
 /**
  * The URL to the Aptos Connect account page if the user is signed in to Aptos Connect.
  *