+++ packages/wallet-adapter-core/src/utils/walletSelector.ts
@@ -1,8 +1,9 @@
-import { WalletInfo } from "./types";
+import { AvailableWallets, WalletInfo } from "./types";
 import { AdapterNotDetectedWallet, AdapterWallet } from "../WalletCore";
 import {
   APTOS_CONNECT_BASE_URL,
   PETRA_WEB_BASE_URL,
+  PETRA_WEB_GENERIC_WALLET_NAME,
   WalletReadyState,
 } from "../constants";
 import { isRedirectable } from "./helpers";
@@ -22,8 +23,11 @@ export function partitionWallets(
   const moreWallets: Array<AdapterNotDetectedWallet> = [];
 
   for (const wallet of wallets) {
-    if (partitionFunction(wallet)) defaultWallets.push(wallet as AdapterWallet);
-    else moreWallets.push(wallet as AdapterNotDetectedWallet);
+    if (partitionFunction(wallet)) {
+      defaultWallets.push(wallet as AdapterWallet);
+    } else {
+      moreWallets.push(wallet as AdapterNotDetectedWallet);
+    }
   }
 
   return { defaultWallets, moreWallets };
@@ -32,7 +36,7 @@ export function partitionWallets(
 /** Returns true if the wallet is installed or loadable. */
 export function isInstalledOrLoadable(
   wallet: AdapterWallet | AdapterNotDetectedWallet,
-) {
+): wallet is AdapterWallet {
   return wallet.readyState === WalletReadyState.Installed;
 }
 
@@ -47,6 +51,16 @@ export function isInstallRequired(
   const isMobile = !isWalletReady && isRedirectable();
 
   return !isMobile && !isWalletReady;
+}
+
+export function shouldUseFallbackWallet(
+  wallet: AdapterWallet | AdapterNotDetectedWallet,
+): boolean {
+  return (
+    !!wallet.fallbackWallet &&
+    isInstallRequired(wallet) &&
+    isInstalledOrLoadable(wallet.fallbackWallet)
+  );
 }
 
 /** Truncates the provided wallet address at the middle with an ellipsis. */
@@ -64,13 +78,26 @@ export function isAptosConnectWallet(wallet: WalletInfo | AdapterWallet) {
   return isPetraWebWallet(wallet);
 }
 
-/** Returns `true` if the provided wallet is a Petra Web wallet. */
-export function isPetraWebWallet(wallet: WalletInfo | AdapterWallet) {
+/** Returns `true` if the provided wallet is a Petra Web wallet. This will automatically exclude the generic wallet. */
+export function isPetraWebWallet(
+  wallet: WalletInfo | AdapterWallet,
+  ignoreGenericWallet: boolean = true,
+) {
   if (!wallet.url) return false;
+  if (ignoreGenericWallet && isPetraWebGenericWallet(wallet)) {
+    return false;
+  }
   return (
     wallet.url.startsWith(APTOS_CONNECT_BASE_URL) ||
     wallet.url.startsWith(PETRA_WEB_BASE_URL)
   );
+}
+
+/** Returns true if the wallet is a generic wallet. */
+export function isPetraWebGenericWallet(
+  wallet: AdapterWallet | AdapterNotDetectedWallet | WalletInfo,
+) {
+  return wallet.name === PETRA_WEB_GENERIC_WALLET_NAME;
 }
 
 /**
@@ -86,7 +113,10 @@ export function getAptosConnectWallets(
     wallets,
     isAptosConnectWallet,
   );
-  return { aptosConnectWallets: defaultWallets, otherWallets: moreWallets };
+  return {
+    aptosConnectWallets: defaultWallets,
+    otherWallets: moreWallets,
+  };
 }
 
 /**
@@ -100,7 +130,10 @@ export function getPetraWebWallets(
     wallets,
     isPetraWebWallet,
   );
-  return { petraWebWallets: defaultWallets, otherWallets: moreWallets };
+  return {
+    petraWebWallets: defaultWallets,
+    otherWallets: moreWallets,
+  };
 }
 
 export interface WalletSortingOptions {
@@ -122,6 +155,24 @@ export interface WalletSortingOptions {
     a: AdapterWallet | AdapterNotDetectedWallet,
     b: AdapterWallet | AdapterNotDetectedWallet,
   ) => number;
+  /**
+   * A map of wallet names to fallback wallet names.
+   * If a wallet is not installed and has a fallback wallet that IS installed,
+   * the fallback wallet will be attached to the not-installed wallet and it will
+   * be moved to the available wallets list.
+   *
+   * @example
+   * ```ts
+   * // Override the default fallbacks
+   * fallbacks: { connections: { "Petra": "Petra Web", "OtherWallet": "OtherFallback" } }
+   * ```
+   */
+  fallbacks?: {
+    /** A map of wallet names to fallback wallet names. */
+    connections: Record<string | AvailableWallets, string | AvailableWallets>;
+    /** An optional array of wallets that are available but not intended for display. These wallets will only be shown if there is logic to explicitly show them (e.g. `fallbacks`) */
+    additionalFallbackWallets?: ReadonlyArray<AdapterWallet>;
+  };
 }
 
 /**
@@ -134,6 +185,8 @@ export interface WalletSortingOptions {
  *
  * `availableWallets` - Wallets that are currently installed or loadable by the client.
  *
+ * `availableWalletsWithFallbacks` - Wallets that are currently uninstalled that have a fallback wallet.
+ *
  * `installableWallets` - Wallets that are NOT current installed or loadable and
  * require the client to install a browser extension first.
  *
@@ -143,9 +196,51 @@ export function groupAndSortWallets(
   wallets: ReadonlyArray<AdapterWallet | AdapterNotDetectedWallet>,
   options?: WalletSortingOptions,
 ) {
+  const {
+    fallbacks: {
+      connections: fallbackConnections,
+      additionalFallbackWallets,
+    } = {},
+  } = options ?? {};
   const { aptosConnectWallets } = getAptosConnectWallets(wallets);
   const { otherWallets, petraWebWallets } = getPetraWebWallets(wallets);
   const { defaultWallets, moreWallets } = partitionWallets(otherWallets);
+
+  // Attach fallback wallets to not-installed wallets and move them to available wallets if the fallback is installed
+  // TODO: Deprecate this and combine this with `availableWallets` in the next major version.
+  const availableWalletsWithFallbacks: Array<
+    AdapterWallet | AdapterNotDetectedWallet
+  > = [];
+
+  if (fallbackConnections && Object.keys(fallbackConnections).length > 0) {
+    for (let i = moreWallets.length - 1; i >= 0; i--) {
+      const wallet = moreWallets[i];
+      const fallbackName = fallbackConnections[wallet.name];
+
+      if (fallbackName) {
+        const fallbackWallet = [
+          ...wallets,
+          ...(additionalFallbackWallets ?? []),
+        ].find((w) => w.name === fallbackName && isInstalledOrLoadable(w)) as
+          | AdapterWallet
+          | undefined;
+
+        // If we found an installed fallback, attach it and move to available wallets
+        if (fallbackWallet) {
+          const walletWithFallback: AdapterNotDetectedWallet = {
+            ...wallet,
+            fallbackWallet,
+          };
+
+          // Remove from installable wallets
+          moreWallets.splice(i, 1);
+
+          // Add to the list to be added to available wallets
+          availableWalletsWithFallbacks.push(walletWithFallback);
+        }
+      }
+    }
+  }
 
   if (options?.sortAptosConnectWallets) {
     aptosConnectWallets.sort(options.sortAptosConnectWallets);
@@ -167,6 +262,8 @@ export function groupAndSortWallets(
     petraWebWallets,
     /** Wallets that are currently installed or loadable. */
     availableWallets: defaultWallets,
+    /** Wallets that are currently uninstalled that have a fallback wallet. */
+    availableWalletsWithFallbacks,
     /** Wallets that are NOT currently installed or loadable. */
     installableWallets: moreWallets,
   };