+++ packages/wallet-adapter-core/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @aptos-labs/wallet-adapter-core
 
+## 8.2.0
+
+### Minor Changes
+
+- 494adef: Add Sui cross-chain wallet support
+
+## 8.1.0
+
+### Minor Changes
+
+- f70e69b: Revert cross-chain wallet initialization in AptosWalletAdapterProvider
+
+## 8.0.1
+
+### Patch Changes
+
+- 7f28e7b: [Fix] Build packages for release
+
+## 8.0.0
+
+### Major Changes
+
+- 378146b: Moved cross-chain wallet initialization into the AptosWalletAdapterProvider
+
+### Minor Changes
+
+- 69c5d26: Added `hideWallets` to `WalletCore` which will allow users optionally hide certain wallets from the default lists of wallets but will be available through the `hiddenWallets` property.
+
+  Also added `fallbacks` to `groupAndSortWallets` which will allow users to specify fallback wallets for wallets that are not installed. This is used to support alternative connection methods for installed wallets, wallets with fallbacks will be moved to a new `availableWalletsWithFallbacks` list.
+
 ## 7.10.2
 
 ### Patch Changes